requirements.txt
h5py
scipy
pandas
tqdm
pillow
scikit-learn
faiss-gpu
matplotlib
torch
tensorboard
timm
opencv-python